GORA-276: Refactor and organize o.a.g.mapreduce package naming
Update pom.xml
