+++ gora-aerospike/src/test/java/org/apache/gora/aerospike/mapreduce/TestAerospikeStoreMapReduceSerialization.java
@@ -19,7 +19,7 @@ package org.apache.gora.aerospike.mapreduce;
 import org.apache.gora.aerospike.store.AerospikeStartupLogWaitStrategy;
 import org.apache.gora.aerospike.store.AerospikeStore;
 import org.apache.gora.examples.generated.WebPage;
-import org.apache.gora.mapreduce.MapReduceTestUtils;
+import org.apache.gora.io.serializer.MapReduceTestUtils;
 import org.apache.gora.store.DataStoreFactory;
 import org.apache.hadoop.conf.Configuration;
 import org.junit.After;
