+++ gora-core/src/examples/java/org/apache/gora/examples/mapreduce/WordCount.java
@@ -24,8 +24,8 @@ import java.util.StringTokenizer;
 
 import org.apache.gora.examples.generated.TokenDatum;
 import org.apache.gora.examples.generated.WebPage;
-import org.apache.gora.mapreduce.GoraMapper;
-import org.apache.gora.mapreduce.GoraReducer;
+import org.apache.gora.io.serializer.GoraMapper;
+import org.apache.gora.io.serializer.GoraReducer;
 import org.apache.gora.query.Query;
 import org.apache.gora.store.DataStore;
 import org.apache.gora.store.DataStoreFactory;
