+++ gora-core/src/main/java/org/apache/gora/flink/GoraFlinkEngine.java
@@ -25,8 +25,8 @@ import org.apache.flink.api.java.hadoop.mapreduce.HadoopOutputFormat;
 import org.apache.flink.api.java.operators.DataSource;
 import org.apache.flink.api.java.tuple.Tuple2;
 import org.apache.flink.util.Preconditions;
-import org.apache.gora.mapreduce.GoraInputFormat;
-import org.apache.gora.mapreduce.GoraOutputFormat;
+import org.apache.gora.io.serializer.GoraInputFormat;
+import org.apache.gora.io.serializer.GoraOutputFormat;
 import org.apache.gora.persistency.impl.PersistentBase;
 import org.apache.gora.store.DataStore;
 import org.apache.gora.store.DataStoreFactory;
