+++ gora-core/src/test/java/org/apache/gora/avro/mapreduce/TestDataFileAvroStoreMapReduce.java
@@ -24,7 +24,7 @@ import java.io.IOException;
 
 import org.apache.gora.avro.store.DataFileAvroStore;
 import org.apache.gora.examples.generated.WebPage;
-import org.apache.gora.mapreduce.DataStoreMapReduceTestBase;
+import org.apache.gora.io.serializer.DataStoreMapReduceTestBase;
 import org.apache.gora.store.DataStore;
 import org.apache.gora.store.DataStoreFactory;
