+++ gora-core/src/test/java/org/apache/gora/io/serializer/package-info.java
@@ -17,4 +17,4 @@
 /**
  * This package contains test cases releated mapreduce.
  */
-package org.apache.gora.mapreduce;
+package org.apache.gora.io.serializer;
