+++ gora-goraci/src/main/java/org/apache/gora/goraci/Verify.java
@@ -31,7 +31,7 @@ import org.apache.commons.cli.GnuParser;
 import org.apache.commons.cli.HelpFormatter;
 import org.apache.commons.cli.Options;
 import org.apache.commons.cli.ParseException;
-import org.apache.gora.mapreduce.GoraMapper;
+import org.apache.gora.io.serializer.GoraMapper;
 import org.apache.gora.query.Query;
 import org.apache.gora.query.Result;
 import org.apache.gora.store.DataStore;
