+++ gora-pig/src/main/java/org/apache/gora/pig/GoraStorage.java
@@ -30,9 +30,9 @@ import org.apache.avro.Schema;
 import org.apache.avro.Schema.Field;
 import org.apache.avro.Schema.Type;
 import org.apache.avro.generic.GenericData.Array;
-import org.apache.gora.mapreduce.GoraMapReduceUtils;
-import org.apache.gora.mapreduce.GoraRecordReader;
-import org.apache.gora.mapreduce.GoraRecordWriter;
+import org.apache.gora.io.serializer.GoraMapReduceUtils;
+import org.apache.gora.io.serializer.GoraRecordReader;
+import org.apache.gora.io.serializer.GoraRecordWriter;
 import org.apache.gora.persistency.impl.PersistentBase;
 import org.apache.gora.pig.mapreduce.PigGoraInputFormat;
 import org.apache.gora.pig.mapreduce.PigGoraOutputFormat;
