+++ gora-tutorial/src/main/java/org/apache/gora/tutorial/log/LogAnalytics.java
@@ -22,8 +22,8 @@ import java.io.IOException;
 import org.apache.avro.util.Utf8;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.apache.gora.mapreduce.GoraMapper;
-import org.apache.gora.mapreduce.GoraReducer;
+import org.apache.gora.io.serializer.GoraMapper;
+import org.apache.gora.io.serializer.GoraReducer;
 import org.apache.gora.store.DataStore;
 import org.apache.gora.store.DataStoreFactory;
 import org.apache.gora.tutorial.log.generated.MetricDatum;
